requirements.txt
deepgram-sdk==3.0.3
Flask==3.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0

openai~=1.13.3